PINS are not more than 255

+++ MK/module/MK_Main.cpp
@@ -10115,7 +10115,7 @@ void kill(const char* lcd_msg) {
 
 #if ENABLED(FAST_PWM_FAN) || ENABLED(FAST_PWM_COOLER)
 
-  void setPwmFrequency(uint8_t pin, int val) {
+  void setPwmFrequency(uint8_t pin, uint8_t val) {
     val &= 0x07;
     switch(digitalPinToTimer(pin)) {
